Move the slip encoder into the flasher module
