+++ espflash/src/lib.rs
@@ -13,7 +13,6 @@ pub mod cli;
 pub mod command;
 pub mod connection;
 pub mod elf;
-pub mod encoder;
 pub mod error;
 pub mod flash_target;
 pub mod flasher;
